Minor API improvements and documentation
Changed the `sparqlEndpoint` method to be an extension function of `Route` instead of `Application`, as otherwise nested paths would not be possible
Reversed the names `SparqlUpdateRequestBuilder::delete` and `SparqlUpdateRequestBuilder::remove` to be more in line with the DSL using RDF-like syntax and non-DSL methods using Kotlin naming
Added various new READMEs to describe the functionality, purpose and usage instructions for the new endpoint modules
