+++ sparql/endpoint/ktor/README.md
@@ -1,0 +1,4 @@
+# SPARQL endpoint Ktor integration
+## Module overview
+The client and server modules offer Ktor integration for using and hosting SPARQL endpoints in your project,
+respectively. Both modules contain an overview of their respective APIs.
